cassandra-backend/requirements.txt
isort
black
mypy
ruff
flask
flask-cors
llama-cpp-python
git+https://github.com/huggingface/transformers.git